moved merge nodes to the TableSelection
